refactor: move adoption tests to entity-specific test files
